CASMINST-2572: add debugging and extend update timeout
In certain situations it appears that the k8s api may not return
the current config map contents if another product is updating
the config map in the meantime. Increase the timeout to allow for
the changes to be available in the API. Also include more logging
for better debugging in these circumstances.