CRAYSAT-1031: Add setuptools installation
This commits adds a setup.py file to cray-product-catalog so that
the script can be installed using setuptools. This also makes it
possible to specify cray-product-catalog as a dependency of another
setuptools-based product.

Note: the setup.py script refers to a non-existent '.version' file
to get the package's version. This must be created manually.

Second, this commit changes the Dockerfile to use the setuptools-based
install instead of copying the catalog_update.py script into the
container image. However, the Dockerfile will link
/usr/bin/catalog_update to /catalog_update.py as this is still the
expected location specified in the cray-import-config base Helm chart.

Finally, add a .gitignore file with a few setuptools artifacts
specified.