CRAYSAT-1031: Add support for removing a product from the catalog
This commit adds catalog_delete.py which a script to delete from the
cray-product-catalog Kubernetes config map. It takes the following
environment variables to control what to remove:

* PRODUCT: Delete the specified product (required)
* PRODUCT_VERSION: Delete the specified product version (required)
* KEY: Within the given product/version, delete the data under KEY
  (optional: if not given, the whole product/version is removed).

Next, this commit adds the catalog_delete.py script to the Docker image.

This commit increments the version to 0.5.0.

Test Description: I ran this script in a Virtualenv on thanos-ncn-m001
to remove product versions and keys within a product version.