CRAYSAT-1101: Set active version
This commit modifies catalog_update.py so that a key of 'active'
is set under $PRODUCT_VERSION with a boolean value whenever running
catalog_update.py. As a part of this, all other versions of $PRODUCT
have their 'active' key set to False.

To opt out, an environment variable called SKIP_SET_ACTIVE_VERSION is
added.

Also, a minor typo of the word 'Kubernetes' is fixed, and unnecessary
'pass' statements are removed.

Test Description:
* Running install.sh while setting CRAY_PRODUCT_CATALOG_UPDATE_IMAGE to
  this version of the image, then inspected the product catalog. I also
  modified the loftsman manifest for the sat-config-import job, so that
  it would also use this version of the cray-product-catalog-update
  image.